REFACTOR: indicio.utils.__init__.py to multiple utils modules
