+++ indicoio/images/fer.py
@@ -1,6 +1,7 @@
 import requests
 
-from indicoio.utils import api_handler, image_preprocess
+from indicoio.utils.api import api_handler
+from indicoio.utils.image import image_preprocess
 import indicoio.config as config
 
 def fer(image, cloud=None, batch=False, api_key=None, **kwargs):
