+++ indicoio/text/tagging.py
@@ -1,4 +1,4 @@
-from indicoio.utils import api_handler
+from indicoio.utils.api import api_handler
 import indicoio.config as config
 
 def text_tags(text, cloud=None, batch=False, api_key=None, **kwargs):
