+++ indicoio/utils/multi.py
@@ -1,6 +1,7 @@
 from indicoio.config import TEXT_APIS, IMAGE_APIS, API_NAMES
+from indicoio.utils.api import api_handler
+from indicoio.utils.image import image_preprocess
 from indicoio.utils.errors import IndicoError
-from indicoio.utils import api_handler, image_preprocess
 
 
 CLIENT_SERVER_MAP = dict((api, api.strip().replace("_", "").lower()) for api in API_NAMES)
